"generate_groups.py" has been updated

+++ apps/groups/management/commands/generate_groups.py
@@ -1,10 +1,9 @@
-import random
-from copy import deepcopy
 from faker import Faker
 from django.core.management.base import BaseCommand
 
+from apps.groups.factories import GroupFactory
 from apps.users.models import User
-from apps.groups.models import GroupInformation, Group
+from apps.groups.models import Group
 
 fake = Faker()
 
@@ -12,27 +11,31 @@ fake = Faker()
 class Command(BaseCommand):
     help = "Generate groups"
 
+    def add_arguments(self, parser):
+        parser.add_argument(
+            "--clear",
+            action="store_true",
+            help="Delete all group records",
+        )
+
     def handle(self, *args, **options):
-        Group.objects.all().delete()
+        if options["clear"]:
+            Group.objects.all().delete()
 
         users = User.objects.all()
+        size = 0
+        try:
+            size = int(input("How much would you like create groups for each user?\n"))
+            if size <= 0:
+                print("Size must be greater than 0.")
+                return
+        except ValueError:
+            print("The value is incorrect.")
+        else:
+            for user in users:
+                groups = GroupFactory.create_batch(size=size, group_info__owner=user)
+                for group in groups:
+                    invited_users = User.objects.exclude(pk=user.id).order_by("?")[:3]
+                    group.invited_users.set(invited_users)
 
-        i = 0
-        while i < 10:
-            copy_users = deepcopy(users)
-            group_info = GroupInformation.objects.create(
-                owner=random.choice(users),
-                name=f"Group №{i + 1}",
-                description=fake.sentence(),
-            )
-            group = Group.objects.create(
-                group_info=group_info,
-            )
-            for _ in range(5):
-                checked_users = []
-                invited_user = random.choice(copy_users)
-                group.invited_users.add(invited_user)
-                checked_users.append(invited_user.username)
-                copy_users = copy_users.exclude(username__in=checked_users)
-            i += 1
-        print("Create groups")
+        print(f"{size * len(users)} groups have been created.")
